WIP: List revdeps using only available packages
Instead of using all_known_packages to list availability, we can simply
use available packages since we are currently only testing on a single
platform and we try to install the packages to test that the build
works. For unavailable packages, the installation wouldn't work anyway.

This lists less packages than are currently using the opam 3 command
magic incantation, but I have convinced myself that this is the correct
list of packages for our use case.